[GENERAL] - Prep work.
Getting the 32-bit build up and running again, complete with process code and able to run CLI etc.
Aiming to get stdlib, and userspace programs (CLI, BASIC, HELLO) all building in 32/64-bit.

Work to do yet, but checking in the current state as it is deploying successfully in full in 32-bit at this point.

+++ kernel/c/common/include/fat32.h
@@ -15,6 +15,7 @@
  */
 
 
+#include <stdint.h>
 
 #ifndef FAT32_H
 #define FAT32_H
@@ -33,7 +34,7 @@
 
     // up to a maximum of 4 register values can be passed in/out of an SWI using this structure
 	struct _kernel_regs {
-		unsigned int r[ 4 ];
+		uintptr_t r[ 4 ];
 	};
 
 	// macro to invoke the SWI instruction [http://www.ethernut.de/en/documents/arm-inline-asm.html] -> asm(code : output operand list : input operand list : clobber list);
